Change registry value removal behaviour
This commit changes how the registry value name is retrieved so that it
can be deleted. Previously, the value name is stored in the
manifest.json on installation, so that the value name can easily be
referred to when deleting.

However, this fails to account for fonts that is not installed with
fonty, or if the fonty manifest.json was manually rebuilt.

The new behaviour searches the entire font registry key manually to find
which value to delete. There's no noticeable performance degradation,
and the code is now more predictable.